Create an interface class for parameters and errors
